src/requirements.txt
discord.py
python-dotenv
supabase
sqlalchemy
asyncpg
psycopg2